refactor(thermostat-card): use new climate entity data
