+++ src/cards/thermostat-card/controls/thermostat-mode-control.ts
@@ -1,17 +1,16 @@
 import { HomeAssistant } from "custom-card-helpers";
-import { HassEntity } from "home-assistant-js-websocket";
 import { css, CSSResultGroup, html, LitElement, TemplateResult } from "lit";
 import { customElement, property } from "lit/decorators.js";
 import { classMap } from "lit/directives/class-map.js";
 import { styleMap } from "lit/directives/style-map.js";
+import { ClimateEntity, compareClimateHvacModes } from "../../../ha/data/climate";
 import { climateIcon } from "../../../utils/icons/climate-icon";
-import { compareClimateHvacModes } from "../utils";
 
 @customElement("mushroom-thermostat-mode-control")
 export class ThermostatModeControl extends LitElement {
     @property({ attribute: false }) public hass!: HomeAssistant;
 
-    @property({ attribute: false }) public entity!: HassEntity;
+    @property({ attribute: false }) public entity!: ClimateEntity;
 
     @property({ type: Boolean }) public fill: boolean = false;
