+++ src/cards/thermostat-card/controls/thermostat-temperature-control.ts
@@ -1,17 +1,17 @@
 import { HomeAssistant } from "custom-card-helpers";
-import { HassEntity } from "home-assistant-js-websocket";
 import { css, CSSResultGroup, html, LitElement, PropertyValues, TemplateResult } from "lit";
 import { customElement, property, state } from "lit/decorators.js";
 import { styleMap } from "lit/directives/style-map.js";
+import { ClimateEntity } from "../../../ha/data/climate";
+import { isActive } from "../../../ha/data/entity";
 import "../../../shared/slider";
-import { isActive } from "../../../utils/entity";
 import { formatDegrees, getStepSize, getTargetTemps } from "../utils";
 
 @customElement("mushroom-thermostat-temperature-control")
 export class ThermostatTemperatureControl extends LitElement {
     @property({ attribute: false }) public hass!: HomeAssistant;
 
-    @property({ attribute: false }) public entity!: HassEntity;
+    @property({ attribute: false }) public entity!: ClimateEntity;
 
     @property({ type: Boolean }) public showIndicators = false;
