--- src/cards/thermostat-card/types.ts
@@ -1,3 +1,0 @@
-export type HvacAction = "off" | "heating" | "cooling" | "drying" | "idle" | "fan";
-
-export type HvacMode = "off" | "heat" | "cool" | "heat_cool" | "auto" | "dry" | "fan_only";
